Rename NotFound.svelte to notfound.svelte to match route filename convention
Agent-Logs-Url: https://github.com/gjelbrim/shqipee/sessions/07e8543e-541a-4384-8ca0-dc9c98176103
